+++ internal/app/interfaces/gateway/device/correlator/handler.go
@@ -1,11 +1,11 @@
-package device
+package correlator
 
 import (
 	"context"
 )
 
-// FGHandler is the interface that describe the FG handler.
-type FGHandler interface {
+// Handler is the interface that describe the correlator handler.
+type Handler interface {
 	Start(ctx context.Context) error
 	Halt(ctx context.Context) error
 	Initialize(ctx context.Context) error
