+++ internal/app/interfaces/gateway/device/fg/handler.go
@@ -1,11 +1,11 @@
-package device
+package fg
 
 import (
 	"context"
 )
 
-// CorrelatorHandler is the interface that describe the correlator handler.
-type CorrelatorHandler interface {
+// Handler is the interface that describe the FG handler.
+type Handler interface {
 	Start(ctx context.Context) error
 	Halt(ctx context.Context) error
 	Initialize(ctx context.Context) error
